Refactor news-specific functions into separate php partial
